+++ src/config/chain/terra.ts
@@ -47,10 +47,6 @@ export const TERRA_CHAIN_CONFIGS: ChainConfigs = {
       gasAdjustment: 1.5,
       maxGasLimit: 25_000_000,
     },
-    explorerLink: {
-      validator: "https://finder.terra.money/mainnet/validator",
-      proposal: "https://station.terra.money/proposal/phoenix-1",
-    },
     extra: {
       singleStakingDenom: "uluna",
     },
@@ -97,10 +93,6 @@ export const TERRA_CHAIN_CONFIGS: ChainConfigs = {
       },
       gasAdjustment: 1.5,
       maxGasLimit: 25_000_000,
-    },
-    explorerLink: {
-      validator: "https://finder.terra.money/testnet/validator",
-      proposal: "https://station.terra.money/proposal/pisco-1",
     },
     extra: {
       singleStakingDenom: "uluna",
