+++ src/lib/components/ExplorerLink.tsx
@@ -1,11 +1,9 @@
 import type { BoxProps, TextProps } from "@chakra-ui/react";
 import { Box, Text, Flex } from "@chakra-ui/react";
 
-import type { ExplorerConfig } from "config/chain/types";
 import { trackMintScan } from "lib/amplitude";
 import { type AddressReturnType } from "lib/app-provider";
 import { useCelatoneApp } from "lib/app-provider/contexts";
-import { useBaseApiRoute } from "lib/app-provider/hooks/useBaseApiRoute";
 import { useWasmConfig } from "lib/app-provider/hooks/useConfig";
 import { useCurrentChain } from "lib/app-provider/hooks/useCurrentChain";
 import { useMobile } from "lib/app-provider/hooks/useMediaQuery";
@@ -44,9 +42,7 @@ export const getNavigationUrl = ({
   wasmEnabled = false,
 }: {
   type: ExplorerLinkProps["type"];
-  explorerConfig: ExplorerConfig;
   value: string;
-  lcdEndpoint: string;
   wasmEnabled?: boolean;
 }) => {
   let url = "";
@@ -172,11 +168,7 @@ export const ExplorerLink = ({
   ...componentProps
 }: ExplorerLinkProps) => {
   const { address } = useCurrentChain();
-  const lcdEndpoint = useBaseApiRoute("rest");
   const { enabled: wasmEnabled } = useWasmConfig({ shouldRedirect: false });
-  const {
-    chainConfig: { explorerLink: explorerConfig },
-  } = useCelatoneApp();
 
   const isInternal =
     type === "code_id" ||
@@ -191,9 +183,7 @@ export const ExplorerLink = ({
   const [hrefLink, textValue] = [
     getNavigationUrl({
       type,
-      explorerConfig,
       value: copyValue || value,
-      lcdEndpoint,
       wasmEnabled,
     }),
     getValueText(value === address, textFormat === "truncate", value),
