+++ src/lib/pages/validator-details/components/bonded-token-changes/index.tsx
@@ -1,10 +1,10 @@
 import { Flex } from "@chakra-ui/react";
 
-import { RelatedTransactionTable } from "../table/RelatedTransactionsTable";
+import { RelatedTransactionTable } from "../tables/RelatedTransactionsTable";
 
 import { VotingPowerChart } from "./VotingPowerChart";
 
-export const BondedTokenChangesSection = () => {
+export const BondedTokenChanges = () => {
   return (
     <Flex direction="column" gap={{ base: 4, md: 8 }} pt={6}>
       <VotingPowerChart denom="OSMO" />
