+++ src/lib/pages/validator-details/components/index.ts
@@ -1,0 +1,5 @@
+export * from "./bonded-token-changes";
+export * from "./performance";
+export * from "./tables";
+export * from "./validator-overview";
+export * from "./validator-top";
