+++ src/lib/pages/validator-details/components/performance/index.tsx
@@ -1,12 +1,12 @@
 import { Flex } from "@chakra-ui/react";
 
-import { ProposedBlocksTable } from "../table/ProposedBlocksTable";
+import { ProposedBlocksTable } from "../tables/ProposedBlocksTable";
 
 import { PenaltySection } from "./PenaltySection";
 import { RecentBlocksSection } from "./RecentBlocksSection";
 import { UptimeSection } from "./UptimeSection";
 
-export const PerformanceSection = () => {
+export const Performance = () => {
   return (
     <Flex direction="column" gap={{ base: 4, md: 6 }} pt={6}>
       <Flex gap={{ base: 4, md: 6 }} direction={{ base: "column", md: "row" }}>
