+++ src/lib/pages/validator-details/components/tables/ProposedBlocksTable.tsx
