+++ src/lib/pages/validator-details/components/tables/index.ts
@@ -1,0 +1,3 @@
+export * from "./ProposedBlocksTable";
+export * from "./RelatedTransactionsTable";
+export * from "./VotedProposalsTable";
