+++ src/lib/pages/validator-details/components/validator-overview/VotingPowerOverview.tsx
@@ -39,6 +39,7 @@ const VotingPowerDetail = ({
     </Text>
   </Flex>
 );
+
 export const VotingPowerOverview = () => {
   return (
     <Flex
